- Refactor Azure/Graph services for modularity & testability - removed AzureManagementService - Update configuration handling, refactor HealthCheckHelper, - revised tests and Azure Functions to use new abstractions. - Improve error handling and code clarity throughout.
