+++ Portal/src/Datahub.Portal/Startup.cs
@@ -207,8 +207,6 @@ public class Startup
         services.AddScoped<ICultureService, UserCultureService>();
 
         services.AddSingleton<IAzureServicePrincipalConfig, AzureServicePrincipalConfig>();
-        services.AddSingleton<AzureManagementService>();
-        services.AddSingleton<ProjectUsageService>();
         services.AddScoped<ProjectStorageConfigurationService>();
 
         //https://github.com/jsakamoto/Toolbelt.Blazor.LocalTimeText/
